Fix: Created utility folder for helper functions.
