+++ components/react/address-card.tsx
@@ -11,7 +11,7 @@ import { WalletStatus } from "@cosmos-kit/core";
 import { FaCheckCircle } from 'react-icons/fa';
 import { FiCopy } from 'react-icons/fi';
 import React, { ReactNode, useEffect,useState } from "react";
-
+import { stringTruncateFromCenter } from "../../utility";
 import { CopyAddressType } from "../types";
 
 const SIZES = {
@@ -35,19 +35,7 @@ const SIZES = {
   },
 };
 
-export function stringTruncateFromCenter(str: string, maxLength: number) {
-  const midChar = '…'; // character to insert into the center of the result
 
-  if (str.length <= maxLength) return str;
-
-  // length of beginning part
-  const left = Math.ceil(maxLength / 2);
-
-  // start index of ending part
-  const right = str.length - Math.floor(maxLength / 2) + 1;
-
-  return str.substring(0, left) + midChar + str.substring(right);
-}
 
 export function handleChangeColorModeValue(
   colorMode: string,
