+++ utility/stringTruncate.tsx
@@ -1,0 +1,13 @@
+export function stringTruncateFromCenter(str: string, maxLength: number) {
+    const midChar = '…'; // character to insert into the center of the result
+  
+    if (str.length <= maxLength) return str;
+  
+    // length of beginning part
+    const left = Math.ceil(maxLength / 2);
+  
+    // start index of ending part
+    const right = str.length - Math.floor(maxLength / 2) + 1;
+  
+    return str.substring(0, left) + midChar + str.substring(right);
+  }
